Revert "feat(reexecution/c): add metervm metrics (#4369)"
This reverts commit 613219c08d126cde462a42741dacab6aaaeb094a.
